+++ udata/tasks.py
@@ -18,6 +18,9 @@ class ContextTask(Task):
     abstract = True
     schedulable = None
     current_app = None
+    route = None
+    default_queue = None
+    default_routing_key = None
 
     def __call__(self, *args, **kwargs):
         with self.current_app.app_context():
@@ -26,6 +29,7 @@ class ContextTask(Task):
 
 class JobTask(ContextTask):
     abstract = True
+    default_queue = 'low'
 
     @property
     def log(self):
@@ -43,14 +47,40 @@ class Scheduler(MongoScheduler):
 celery = Celery(task_cls=ContextTask)
 
 
+def router(name, args, kwargs, options, task=None, **kw):
+    '''
+    A celery router using the predeclared :class:`ContextTask`
+    attributes (`router` or `default_queue` and/or `default routing_key`).
+    '''
+    # Fetch task by name if necessary
+    task = task or celery.tasks.get(name)
+    if not task:
+        return
+    # Single route param override everything
+    if task.route:
+        queue = task.route.split('.', 1)[0]
+        return {'queue': queue, 'routing_key': task.route}
+    # queue parameter, routing_key computed if not present
+    if task.default_queue:
+        key = task.default_routing_key
+        key = key or '{0.default_queue}.{0.name}'.format(task)
+        return {'queue': task.default_queue, 'routing_key': key}
+    # only routing_key, queue should not be returned to fallback on default
+    elif task.default_routing_key:
+        return {'routing_key': task.default_routing_key}
+
+
 def task(*args, **kwargs):
+    for arg in 'queue', 'routing_key':
+        if arg in kwargs:
+            kwargs['default_{0}'.format(arg)] = kwargs.pop(arg)
     return celery.task(*args, **kwargs)
 
 
 def job(name, **kwargs):
     '''A shortcut decorator for declaring jobs'''
-    return celery.task(name=name, schedulable=True, base=JobTask,
-                       bind=True, **kwargs)
+    return task(name=name, schedulable=True, base=JobTask,
+                bind=True, **kwargs)
 
 
 def get_logger(name):
@@ -76,6 +106,19 @@ def log_test(self):
     self.log.info('This is an INFO message')
     self.log.warning('This is a WARNING message')
     self.log.error('This is a ERROR message')
+
+
+@job('test-low-queue', queue='low')
+@job('test-default-queue', queue='default')
+@job('test-high-queue', queue='high')
+def queue_test(self, raises=False):
+    '''
+    An empty task to test the queues.
+    Set raise to True if you want a notification (ie. Sentry)
+    '''
+    self.log.info('Just a test task')
+    if raises:
+        raise Exception('Test task processing')
 
 
 @job('error-test')