+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - udata is now protocol (`http`/`https`) agnostic. This is now fully the reverse-proxy responsibility (please ensure that you are using SSL only in production for security purpose). [#1463](https://github.com/opendatateam/udata/pull/1463) [breaking]
 - Fix organization widget (embed) [#1474](https://github.com/opendatateam/udata/pull/1474)
 - High priority for sendmail tasks [#1484](https://github.com/opendatateam/udata/pull/1484)
+- Improve tasks/jobs queues routing [#1487](https://github.com/opendatateam/udata/pull/1487)
 
 ## 1.2.11 (2018-02-05)